pyproject.toml
[project]
name = "meta"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dedalus-labs>=0.1.0a9",
    "flask==3.0.0",
    "flask-cors==4.0.0",
    "flask-socketio>=5.5.1",
    "pyngrok==7.0.0",
    "python-dotenv>=1.1.1",
    "requests==2.31.0",
]
